change test name to PR 1411
